+++ src/lib/reasoners/sat_solver_sig.mli
@@ -49,7 +49,7 @@ exception Sat
 exception Unsat of Explanation.t
 exception I_dont_know
 
-module type S = sig
+module type Internal = sig
   type t
 
   val empty : ?selector:(Expr.t -> bool) -> unit -> t
@@ -59,7 +59,7 @@ module type S = sig
       discovered by the instantiation engine. *)
 
   val declare : t -> Symbols.typed_name -> unit
-  (** [declare env nmae] declares a new typed name [name].
+  (** [declare env name] declares a new typed name [name].
 
       If the environment [env] isn't unsatisfiable and the model generation
       is enabled, the solver produces a model term for [name] which can be
@@ -117,7 +117,13 @@ module type S = sig
   val reinit_ctx : unit -> unit
   (** [reinit_ctx ()] reinitializes the solving context. *)
 
-  val get_model: t -> Models.t option
+  val get_boolean_model : t -> Expr.t list option
+  (** [get_boolean_model env] returns the list of literals assigned by
+      the solver [env].
+
+      @return [None] if we don't perform any [assume]. *)
+
+  val get_model : t -> Models.t option
   (** [get_model t] produces the current first-order model.
       Notice that this model is a best-effort.
 
@@ -128,10 +134,6 @@ module type S = sig
   (** [get_unknown_reason t] returns the reason Alt-Ergo raised
       [I_dont_know] if it did. If it did not, returns [None]. *)
 
-  val get_value : t -> Expr.t -> Expr.t option
-  (** [get_value t e] returns the value of [e] as a constant expression
-      in the current model generated. Returns [None] if can't decide. *)
-
   val get_objectives : t -> Objective.Model.t option
   (** [get_objectives t] returns the current objective values.
 
@@ -139,6 +141,31 @@ module type S = sig
       unsatisfiable. *)
 end
 
+type lbool = True | False | Unknown
+
+val pp_lbool : lbool Fmt.t
+
+module type S = sig
+  include Internal
+
+  val get_value : t -> Expr.t list -> Expr.t list option
+  (** [get_value env l] returns the model values of the expressions of [l]
+      in the current generated model of [env].
+
+      @return [None] if the model generation is not enabled or the
+              environment is already unsatisfiable before calling this function.
+      @raise Unsat if the solver found a contradiction. *)
+
+  val get_assignment : t -> Expr.t list -> lbool list
+  (** [get_assignment env l] returns the status of the literals [l] in the
+      current boolean model of [env].
+
+      The status is [unknown] if the literal isn't a subformula of the user
+      input.
+
+      @raise invalid_argument if one of the expressions of [l] isn't a
+             literal. *)
+end
 
 module type SatContainer = sig
   module Make (Th : Theory.S) : S